# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "React Online Store",
        "version": "0.1"
    },
    "paths": {
        "/api/admin/brands": {
            "get": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Display a listing of the Brands.",
                "operationId": "75605c4db2e3dc0225481241d0d7921a",
                "responses": {
                    "200": {
                        "description": "Display a listing of the Brands",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BrandResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Store a newly created Brand.",
                "operationId": "4f025f26cd6259684ebc61a96ca06828",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBrandRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stored a newly created Brand",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/brands/{id}": {
            "get": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Display the specified Brand by ID.",
                "operationId": "a6da6c39276d50ac8382d21fc4f184d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Brand by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Update the specified Brand by ID.",
                "operationId": "0d2cf120fdb808ee402c9ba1874e10fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBrandRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated the specified Brand by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Remove the specified Brand by ID.",
                "operationId": "a5b037cfc0c0421d256deb3334ff71b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Brand by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/brands/{id}/restore": {
            "post": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Restore the specified Brand by ID.",
                "operationId": "6692ecc0ba05270a45419300aed2a34a",
                "responses": {
                    "200": {
                        "description": "Restore brand"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/brands/{id}/force": {
            "delete": {
                "tags": [
                    "AdminBrand"
                ],
                "summary": "Remove the specified Brand by ID.",
                "operationId": "2ca731be77ebfdc148267c275bbb953b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Brand by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/brands": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Display a listing of the Brands.",
                "operationId": "130f2bd423b6fe2355253e2c39878d99",
                "responses": {
                    "200": {
                        "description": "Display a listing of the Brands",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BrandResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/brands/{id}": {
            "get": {
                "tags": [
                    "Brand"
                ],
                "summary": "Display the specified Brand by ID.",
                "operationId": "2086e16f9b1a70aea714677ef42808a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Brand by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BrandResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreBrandRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Brand name",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "UpdateBrandRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Brand name",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
            "BrandResource": {
                "properties": {
                    "id": {
                        "description": "Class BrandResource",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
...
            "ImageResource": {
                "properties": {
                    "id": {
                        "description": "Class ImageResource",
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "origin_name": {
                        "type": "string"
                    },
                    "width": {
                        "type": "string"
                    },
                    "height": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    }
}